chore: reorganize component strcutures
