+++ src/components/Custom/DatePicker/DatePicker.tsx
@@ -1,8 +1,10 @@
-import MultiDatePicker from "react-multi-date-picker";
+import MultiDatePicker, { DateObject } from "react-multi-date-picker";
 import persian from "react-date-object/calendars/persian";
 import persian_fa from "react-date-object/locales/persian_fa";
 import { cva, type VariantProps } from "class-variance-authority";
 import "./DatePicker.css";
+import { useField } from "formik";
+import { toTehranISOString } from "@/utils/toTehranISOString";
 
 const inputVariants = cva(
 	"flex h-13 w-full !text-[15px] rounded-full border border-[1px] border-black/40 bg-white font-[Alibaba] font-bold px-6 py-1 text-base transition-colors file:border-0 file:bg-transparent file:text-sm file:font-medium file:text-foreground placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:cursor-not-allowed disabled:opacity-50 md:text-sm",
@@ -18,23 +20,40 @@ const inputVariants = cva(
 		},
 	},
 );
-export interface InputProps
+export interface DatePickerProps
 	extends Omit<React.InputHTMLAttributes<HTMLInputElement>, "className">,
 		VariantProps<typeof inputVariants> {
 	asChild?: boolean;
 }
-export default function DatePicker({ shadow, ...props }: InputProps) {
+export default function DatePicker({
+	shadow,
+	name,
+	...props
+}: DatePickerProps) {
+	const [field, , helpers] = useField(name || "");
 	return (
 		<div>
 			<MultiDatePicker
+				value={
+					field.value
+						? new DateObject({
+								date: new Date(field.value),
+								calendar: persian,
+								locale: persian_fa,
+							})
+						: ""
+				}
+				onChange={(value) => {
+					helpers.setValue(toTehranISOString(value!.toDate()));
+				}}
 				calendar={persian}
 				locale={persian_fa}
 				inputClass={inputVariants({ shadow })}
-				mapDays={({ date, today }: MapDaysArgs): Partial<DayValue> => {
-					const props: Partial<DayValue> = {};
+				mapDays={({ date }) => {
+					const props: Partial<{ className?: string }> = {};
 
 					if (date.weekDay.index === 6) {
-						props.className = "text-primary";
+						props.className = "text-red-500";
 					}
 
 					return props;
