+++ src/pages/Test.tsx
@@ -15,8 +15,8 @@ import {
 	SelectTrigger,
 	SelectValue,
 } from "@/components/Custom/Select/Select";
-import { PetDatePicker } from "@/components/Custom/DatePicker/PetDatePicker";
 import DatePicker from "@/components/Custom/DatePicker/DatePicker";
+import { PetDatePicker } from "@/components/Custom/PetDatePicker/PetDatePicker";
 
 const validationSchema = Yup.object({
 	email: Yup.string()
@@ -55,7 +55,7 @@ function Test() {
 								placeholder="ایمیل"
 							/>
 						</div>
-						<DatePicker />
+						<DatePicker name="akhoond2" />
 
 						<div className="mt-5 w-50">
 							<Input
